shimmy types around for traits
