+++ .vscode/settings.json
@@ -13,5 +13,7 @@
   ],
   "[terraform]": {
     "editor.formatOnSave": false
-  }
+  },
+  "javascript.updateImportsOnFileMove.enabled": "always",
+  "typescript.updateImportsOnFileMove.enabled": "always"
 }
