+++ packages/api/src/models/Trait/gql-types/TraitFilters.ts
@@ -1,0 +1,11 @@
+import { IsUUID } from "class-validator";
+import { InputType, Field, ID } from "type-graphql";
+
+@InputType()
+export class TraitFilters {
+  @Field(() => ID, {
+    nullable: false,
+  })
+  @IsUUID(4)
+  speciesId!: string;
+}
