+++ packages/api/src/models/Trait/gql-types/TraitListResponse.ts
@@ -1,0 +1,6 @@
+import { createUnionType } from "type-graphql";
+
+export const TraitListResponse = createUnionType({
+  name: "TraitListResponse",
+  types: () => [],
+});
